+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `SelectStack` component
+
 ### Fixed
 
 ## [9.5.0] 2023-04-12